Function sentry_core::start_transaction

pub fn start_transaction(ctx: TransactionContext) -> Transaction

Start a new Performance Monitoring Transaction.

The transaction needs to be explicitly finished via Transaction::finish, otherwise it will be discarded. The transaction itself also represents the root span in the span hierarchy. Child spans can be started with the Transaction::start_child method.