Trait sentry_core::FutureExt

pub trait FutureExt: Sized {
    fn bind_hub<H>(self, hub: H) -> SentryFuture<Self>
    where
        H: Into<Arc<Hub>>,
    { ... }
}

Future extensions for Sentry.

Provided methods

fn bind_hub<H>(self, hub: H) -> SentryFuture<Self> where
H: Into<Arc<Hub>>,

Binds a hub to the execution of this future.

This ensures that the future is polled within the given hub.

Loading content...