Module sentry_core::types::protocol::v7::map

The internally used arbitrary data map type.

Structs

BTreeMap
A map based on a B-Tree.

IntoIter
An owning iterator over the entries of a BTreeMap.

Iter
An iterator over the entries of a BTreeMap.

IterMut
A mutable iterator over the entries of a BTreeMap.

Keys
An iterator over the keys of a BTreeMap.

Map
A map based on a B-Tree.

OccupiedEntry
A view into an occupied entry in a BTreeMap. It is part of the Entry enum.

Range
An iterator over a sub-range of entries in a BTreeMap.

RangeMut
A mutable iterator over a sub-range of entries in a BTreeMap.

VacantEntry
A view into a vacant entry in a BTreeMap. It is part of the Entry enum.

Values
An iterator over the values of a BTreeMap.

ValuesMut
A mutable iterator over the values of a BTreeMap.

DrainFilter	Experimental
An iterator produced by calling drain_filter on BTreeMap.

Enums

Entry
A view into a single entry in a map, which may either be vacant or occupied.